# backend/api.yaml
...
openapi: 3.0.3
info:
  title: Katalyst Placement Portal API
  description: |
    Decoupled Flask JSON API for Katalyst V2.

    - Auth: `Authorization: Bearer <access_token>` (JWT)
    - Roles: `admin`, `student`, `company` (blacklisted users get 403)
    - UI wording uses "drives"; API resources use `job_positions`
    - Errors: `{ "error", "message", "status" }`
  version: 2.0.0
  contact:
    name: Deva Vasista
    email: 21f2000226@ds.study.iitm.ac.in
servers:
  - url: http://127.0.0.1:3000
    description: Local Flask API (`cd backend && python app.py`)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/auth/login
    RefreshAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token from login (use only on /api/auth/refresh)
  schemas:
    Error:
      type: object
      required: [error, message, status]
      properties:
        error: { type: string, example: Bad Request }
        message: { type: string }
        status: { type: integer, example: 400 }
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string, format: email }
        role:
          type: string
          enum: [admin, student, company, blacklisted]
        created_at: { type: string, format: date-time }
        student: { $ref: "#/components/schemas/Student" }
        company: { $ref: "#/components/schemas/Company" }
    Student:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        full_name: { type: string }
        phone: { type: string, nullable: true }
        resume_path: { type: string, nullable: true }
        cgpa: { type: number, format: float, nullable: true }
        graduation_year: { type: integer, nullable: true }
        skills: { type: string, nullable: true }
        notify_deadlines: { type: boolean }
        created_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    Company:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        name: { type: string }
        website: { type: string, nullable: true }
        hr_contact: { type: string, nullable: true }
        industry: { type: string, nullable: true }
        approval_status:
          type: string
          enum: [pending, approved, rejected]
        created_at: { type: string, format: date-time }
        user: { $ref: "#/components/schemas/User" }
    JobPosition:
      type: object
      description: Placement drive (UI name)
      properties:
        id: { type: integer }
        company_id: { type: integer }
        title: { type: string }
        description: { type: string, nullable: true }
        requirements: { type: string, nullable: true }
        min_cgpa:
          type: number
          format: float
          nullable: true
          description: NULL means no CGPA cutoff
        salary_min: { type: integer, nullable: true }
        salary_max: { type: integer, nullable: true }
        deadline: { type: string, format: date-time, nullable: true }
        status:
          type: string
          enum: [active, closed]
        approval_status:
          type: string
          enum: [pending, approved, rejected]
        created_at: { type: string, format: date-time }
        company: { $ref: "#/components/schemas/Company" }
    Application:
      type: object
      properties:
        id: { type: integer }
        student_id: { type: integer }
        job_position_id: { type: integer }
        status:
          type: string
          enum: [applied, shortlisted, selected, rejected]
        interview_at: { type: string, format: date-time, nullable: true }
        interview_notes: { type: string, nullable: true }
        applied_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        student: { $ref: "#/components/schemas/Student" }
        job_position: { $ref: "#/components/schemas/JobPosition" }
...
    JobPositionInput:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        requirements: { type: string }
        min_cgpa: { type: number, format: float, nullable: true }
        salary_min: { type: integer, nullable: true }
        salary_max: { type: integer, nullable: true }
        deadline:
          type: string
          description: YYYY-MM-DD or ISO datetime
      required: [title]
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Forbidden:
      description: Wrong role or blacklisted
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/company/dashboard:
    get:
      tags: [Company]
      summary: Company dashboard
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Company, stats, drives by approval bucket
          content:
            application/json:
              schema:
                type: object
                properties:
                  company: { $ref: "#/components/schemas/Company" }
                  stats: { type: object, additionalProperties: true }
                  approved_job_positions:
                    type: array
                    items: { $ref: "#/components/schemas/JobPosition" }
                  pending_job_positions:
                    type: array
                    items: { $ref: "#/components/schemas/JobPosition" }
                  rejected_job_positions:
                    type: array
                    items: { $ref: "#/components/schemas/JobPosition" }
  /api/company/profile:
    get:
      tags: [Company]
      summary: Get company profile
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  company: { $ref: "#/components/schemas/Company" }
    put:
      tags: [Company]
      summary: Update company profile
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                company_name: { type: string }
                website: { type: string }
                hr_contact: { type: string }
                industry: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  company: { $ref: "#/components/schemas/Company" }
  /api/company/job-positions:
    post:
      tags: [Company]
      summary: Create drive (pending admin approval)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobPositionInput" }
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  job_position: { $ref: "#/components/schemas/JobPosition" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/company/job-positions/{job_id}:
    get:
      tags: [Company]
      summary: Own drive detail and applications
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Drive and applicants
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_position: { $ref: "#/components/schemas/JobPosition" }
                  applications:
                    type: array
                    items: { $ref: "#/components/schemas/Application" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Company]
      summary: Edit own drive (re-pends if previously approved)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/JobPositionInput" }
      responses:
        "200":
          description: Updated
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Company]
      summary: Delete own drive
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Deleted
  /api/company/job-positions/{job_id}/toggle:
    post:
      tags: [Company]
      summary: Close active drive, or request reopen (pending approval)
      security: [{ BearerAuth: [] }]
      parameters:
        - name: job_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  job_position: { $ref: "#/components/schemas/JobPosition" }
  /api/company/applications/{application_id}/status:
    patch:
      tags: [Company]
      summary: Shortlist, reject, or select an applicant
      description: |
        Selecting a student creates a Placement (`salary` required, `joining_date` optional).
        Shortlisting may email the student (MailHog locally) when SMTP is configured.
      security: [{ BearerAuth: [] }]
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [applied, shortlisted, selected, rejected]
                interview_at:
                  type: string
                  description: datetime-local / ISO; used when shortlisting
                interview_notes: { type: string }
                salary:
                  type: number
                  description: Required when status is selected (LPA)
                joining_date:
                  type: string
                  format: date
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  emailed: { type: boolean }
                  application: { $ref: "#/components/schemas/Application" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/company/students/{student_id}/resume:
    get:
      tags: [Company]
      summary: Download resume of a student who applied to this company
      security: [{ BearerAuth: [] }]
      parameters:
        - name: student_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Resume file
          content:
            application/pdf: { schema: { type: string, format: binary } }
            application/msword: { schema: { type: string, format: binary } }
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema: { type: string, format: binary }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---------- Student ----------
...